+++ agate/table/__init__.py
@@ -532,6 +532,7 @@ from agate.table.bins import bins
 from agate.table.compute import compute
 from agate.table.denormalize import denormalize
 from agate.table.from_csv import from_csv
+from agate.table.from_fixed import from_fixed
 from agate.table.from_json import from_json
 from agate.table.from_object import from_object
 from agate.table.group_by import group_by
@@ -552,6 +553,7 @@ Table.bins = bins
 Table.compute = compute
 Table.denormalize = denormalize
 Table.from_csv = from_csv
+Table.from_fixed = from_fixed
 Table.from_json = from_json
 Table.from_object = from_object
 Table.group_by = group_by